Struct portaudio::hostapi::HostApiInfo

pub struct HostApiInfo {
    pub api_type: HostApiType,
    pub name: String,
    pub device_count: int,
    pub default_input: Option<int>,
    pub default_output: Option<int>,
}

Information about a specific host API

Fields

api_type

The type of the API

name

Human-readable name of the API

device_count

Number of devices this API has

default_input

Default input device of the API. Is None if there is no input device available.

default_output

Default output device of the API. Is None if there is no output device available.