Module portaudio::pa

The portable PortAudio API.

Modules

device
Device management

host
Host api management

Structs

PaStream
Representation of an audio stream, where the format of the stream is defined by the S parameter.

Functions

get_error_text
Translate the supplied PortAudio error code into a human readable message.

get_last_host_error_info
Return information about the last host error encountered. The error information returned by get_last_host_error_info() will never be modified asynchronously by errors occurring in other PortAudio owned threads (such as the thread that manages the stream callback.)

get_sample_size
Retrieve the size of a given sample format in bytes.

get_version
Retrieve the release number of the currently running PortAudio build.

get_version_text
Retrieve a textual description of the current PortAudio build.

initialize
Library initialization function - call this before using PortAudio. This function initializes internal data structures and prepares underlying host APIs for use. With the exception of get_version(), get_version_text(), and get_error_text(), this function MUST be called before using any other PortAudio API functions.

is_format_supported
Determine whether it would be possible to open a stream with the specified parameters.

sleep
Put the caller to sleep for at least 'msec' milliseconds. This function is provided only as a convenience for authors of portable code (such as the tests and examples in the PortAudio distribution.)

terminate
Library termination function - call this when finished using PortAudio. This function deallocates all resources allocated by PortAudio since it was initialized by a call to initialize(). In cases where initialise() has been called multiple times, each call must be matched with a corresponding call to terminate(). The final matching call to terminate() will automatically close any PortAudio streams that are still open.