Module portaudio::pa::host

Host api management

Functions

api_device_index_to_device_index
Convert a host-API-specific device index to standard PortAudio device index. This function may be used in conjunction with the deviceCount field of PaHostApiInfo to enumerate all devices for the specified host API.

api_type_id_to_host_api_index
Convert a static host API unique identifier, into a runtime host API index.

get_api_count
Retrieve the number of available host APIs. Even if a host API is available it may have no devices available.

get_api_info
Retrieve a pointer to a structure containing information about a specific host Api.

get_default_api
Retrieve the index of the default host API. The default host API will be the lowest common denominator host API on the current platform and is unlikely to provide the best performance.