portaudio

Module stream

This module aims to provide a user-friendly rust-esque wrapper around the portaudio Stream types.

The primary type of interest is Stream.

Re-exports

pub use self::callback_flags::CallbackFlags;

pub use self::flags::Flags;

Modules

callback_flags

A type safe wrapper around PortAudio’s stream callback flags.

flags

A type safe wrapper around PortAudio’s stream flags.

Structs

Blocking

A Stream Mode representing a blocking stream.

Buffer

The buffer used to transfer audio data between the input and output streams.

Duplex

A type of Flow that describes a bi-directional (input and output) Stream.

DuplexCallbackArgs

Arguments given to a NonBlocking Input Stream’s CallbackFn.

DuplexCallbackTimeInfo

Timing information for the buffers passed to the stream callback.

DuplexSettings

Settings used to construct a Duplex Stream.

Info

A structure containing unchanging information about an open stream.

Input

A type of Flow that describes an input-only Stream.

InputCallbackArgs

Arguments given to a NonBlocking Input Stream’s CallbackFn.

InputCallbackTimeInfo

Timing information for the buffer passed to the input stream callback.

InputSettings

Settings used to construct an Input Stream.

NonBlocking

A Stream Mode representing a non-blocking stream.

Output

A type of Flow that describes an output-only Stream.

OutputCallbackArgs

Arguments given to a NonBlocking Input Stream’s CallbackFn.

OutputCallbackTimeInfo

Timing information for the buffer passed to the output stream callback.

OutputSettings

Settings used to construct an Out Stream.

Parameters

Parameters for one direction (input or output) of a stream.

Stream

A type-safe PortAudio PaStream wrapper.

Enums

Available

Describes stream availability and the number for frames available for reading/writing if there is any.

Traits

Flow

There are three possible Flows available for a Stream: Input, Out and Duplex.

Mode

There are two Modes with which a Stream can be set: Blocking and NonBlocking.

Reader

Streams that can be read by the user.

Settings

Types used to open a Stream via the PortAudio::open_blocking_stream and PortAudio::open_non_blocking_stream methods.

Writer

Streams that can be written to by the user for output to some DAC.