portaudio::stream

Struct Stream

pub struct Stream<M, F> { /* private fields */ }

A type-safe PortAudio PaStream wrapper.

F is the stream’s directional Flow:

Input - Receives data from an input device’s ADC.

Output - Sends data to an output device’s DAC.

Duplex - Receives and Sends data on two devices synchronously.

A Stream of a particular Flow type can be opened by passing the Flow’s associated Settings type to either of the PortAudio::open_blocking_stream or PortAudio::open_non_blocking_stream methods.

InputSettings -> Input

OutputSettings -> Output

DuplexSettings -> Duplex

M is the stream’s Mode:

Blocking - The stream will be run on the caller’s thread. For Blocking streams, a user can read from Input and Duplex streams using the Stream::read_available and Stream::read methods and write to Output and Duplex streams using the Stream::write_available and Stream::write methods. A Blocking Stream can be opened using the **PortAudio::open_blocking_stream** method.

NonBlocking - The stream will be run on a separate thread. **NonBlocking streams are read and written to via the callback arguments that are associated with the Stream’s Flow type:

Input -> InputCallbackArgs

Output -> OutputCallbackArgs

Duplex -> DuplexCallbackArgs A NonBlocking Stream can be opened using the **PortAudio::open_non_blocking_stream** method.

A Stream may only live as long as the PortAudio instance from which it was spawned and no longer.

The original PortAudio documentation for the PaStream type can be found here.

Implementations

impl<M, F> Stream<M, F>

pub fn close(&mut self) -> Result<(), Error>

Closes an audio stream.

If the audio stream is active it discards any pending buffers as if Stream::abort had been called.

pub fn start(&mut self) -> Result<(), Error>

Commences audio processing.

pub fn stop(&mut self) -> Result<(), Error>

Terminates audio processing.

It waits until all pending audio buffers have been played before it returns.

pub fn abort(&mut self) -> Result<(), Error>

Terminates audio processing immediately without waiting for pending buffers to complete.

pub fn is_stopped(&self) -> Result<bool, Error>

Determine whether the stream is stopped.

A stream is considered to be stopped prior to a successful call to start_stream and after a successful call to stop_stream or abort_stream.

If a stream callback returns a value other than Continue the stream is NOT considered to be stopped.

Return true when the stream is stopped.

Returns false when the stream is running.

Returnes Error if an error is encountered.

TODO: Clarify what errors can actually an occur.

pub fn is_active(&self) -> Result<bool, Error>

Determine whether the stream is active.

A stream is active after a successful call to Stream::start, until it becomes inactive either as a result of a call to Stream::stop or Stream::abort, or as a result of a return value other than Continue from the stream callback. In the latter case, the stream is considered inactive after the last buffer has finished playing.

Returns true when the stream is active (ie playing or recording audio).

Returns false when not playing.

Returns an Error if an error is encountered.

TODO: Clarify what errors can actually an occur.

pub fn time(&self) -> Time

Returns the current time in seconds for a stream according to the same clock used to generate callback CallbackTimeInfo timestamps.

The time values are monotonically increasing and have unspecified origin.

This returns valid time values for the entire life of the stream, from when the stream is opened until it is closed.

Starting and stopping the stream does not affect the passage of time returned by this method.

Returns the stream’s current time in seconds, or 0 if an error occurred.

pub fn info(&self) -> Info

Retrieve a Info structure containing information about the stream.

pub fn unsafe_pa_stream(&self) -> *mut PaStream

This function is solely for use within the extension modules for interacting with PortAudio platform-specific extension APIs.

impl<F> Stream<Blocking<F::Buffer>, F>
where F: Flow,

pub fn open<S>(life: Arc<Life>, settings: S) -> Result<Self, Error>
where S: Settings<Flow = F>,

Open a new Blocking Stream with the given Flow and settings.

impl<F> Stream<Blocking<F::Buffer>, F>
where F: Flow + Reader,

pub fn read_available(&self) -> Result<Available, Error>

Retrieve the number of frames that can be read from the stream without waiting.

Returns a Result with either:

An Ok variant with a Available enum describing either:

The number of frames available to be read from the stream (without blocking or busy waiting) or

Flags indicating whether or not there has been input overflow or output underflow.

An Err variant in the case PortAudio is not initialized or some error is encountered.

See the blocking.rs example for a usage example.

pub fn read<'b>(&'b self, frames: u32) -> Result<&'b [F::Sample], Error>

Read samples from an input stream.

The function doesn’t return until the entire buffer has been filled - this may involve waiting for the operating system to supply the data.

Arguments

frames - The number of frames in the buffer.

Returns an interleaved slice containing the read audio data.

Returns an Error if some error occurred.

TODO: Research and document exactly what errors can occur.

impl<F> Stream<Blocking<F::Buffer>, F>
where F: Flow + Writer,

pub fn write_available(&self) -> Result<Available, Error>

Retrieve the number of frames that can be written to the stream without waiting.

Returns a Result with either:

An Ok variant with a Available enum describing either:

The number of frames available to be written to the stream (without blocking or busy waiting) or

Flags indicating whether or not there has been input overflow or output underflow.

An Err variant in the case PortAudio is not initialized or some error is encountered.

See the blocking.rs example for a usage example.

pub fn write<WF>(&mut self, frames: u32, write_fn: WF) -> Result<(), Error>
where WF: for<'b> FnOnce(&'b mut [F::Sample]),

Write samples to an output stream.

This function doesn’t return until the entire buffer has been consumed

this may involve waiting for the operating system to consume the data.

Arguments

frames - The number of frames in the buffer.

write_fn - The buffer contains samples in the format specified by S.

Returns Ok(()) on success and an Err(Error) variant on failure.

impl<F> Stream<NonBlocking, F>

pub fn open<S, C>( life: Arc<Life>, settings: S, callback: C, ) -> Result<Self, Error>
where S: Settings<Flow = F>, F: Flow, C: FnMut(F::CallbackArgs) -> PaStreamCallbackResult + 'static,

Open a new NonBlocking Stream with the given Flow and settings.

pub fn cpu_load(&self) -> f64

Retrieve CPU usage information for the specified stream.

The “CPU Load” is a fraction of total CPU time consumed by a callback stream’s audio processing routines including, but not limited to the client supplied stream callback.