portaudio::stream

Trait Flow

pub trait Flow {
    type Buffer;
    type CallbackArgs;
    type CallbackTimeInfo;

    // Required methods
    fn new_buffer(&self, frames_per_buffer: u32) -> Self::Buffer;
    fn params_both_directions(
        &self,
    ) -> (Option<PaStreamParameters>, Option<PaStreamParameters>);
    fn new_callback_args(
        input: *const c_void,
        output: *mut c_void,
        frame_count: c_ulong,
        time_info: *const PaStreamCallbackTimeInfo,
        flags: PaStreamCallbackFlags,
        in_channels: i32,
        out_channels: i32,
    ) -> Self::CallbackArgs;
}

There are three possible Flows available for a Stream: Input, Out and Duplex.

Required Associated Types

type Buffer

The type of buffer(s) necessary for transferring audio in a Blocking stream.

type CallbackArgs

The arguments passed to the non-blocking stream callback.

type CallbackTimeInfo

Timing information for the buffer passed to the stream callback.

Required Methods

fn new_buffer(&self, frames_per_buffer: u32) -> Self::Buffer

Construct a new Self::Buffer.

fn params_both_directions( &self, ) -> (Option<PaStreamParameters>, Option<PaStreamParameters>)

Necessary for dynamically acquiring bi-directional params for Pa_OpenStream.

fn new_callback_args( input: *const c_void, output: *mut c_void, frame_count: c_ulong, time_info: *const PaStreamCallbackTimeInfo, flags: PaStreamCallbackFlags, in_channels: i32, out_channels: i32, ) -> Self::CallbackArgs

Constructs the Flow’s associated CallbackArgs from the non-blocking C API stream parameters.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.