portaudio

Struct PortAudio

pub struct PortAudio { /* private fields */ }

A type-safe wrapper around the PortAudio API.

We use a type here instead of pure functions in order to ensure correct intialisation and termination of the underlying PortAudio instance.

Implementations

impl PortAudio

pub fn new() -> Result<Self, Error>

Construct a PortAudio instance.

This calls PortAudio’s Pa_Initialize function which initializes internal data structures and prepares underlying host APIs for use.

Using the C API, a user would normally have to call Pa_Terminate when shutting down PortAudio, however this PortAudio type will automatically take care of cleanup when Dropped.

It is safe to simultaneously construct more than one PortAudio instance, however this is rarely necessary.

pub fn terminate(self) -> Result<(), Error>

Takes ownership of self and terminates the PortAudio API using Pa_Terminate.

This function deallocates all resources allocated by PortAudio since it was constructed.

Calling this method is optional. It is only necessary if you require handling any PortAudio termination errors. Otherwise, Pa_Terminate will be called and all necessary cleanup will occur automatically when this PortAudio instance is Dropped.

pub fn version(&self) -> i32

Retrieve the release number of the currently running PortAudio build.

pub fn version_text(&self) -> Result<&'static str, Utf8Error>

Retrieve a textual description of the current PortAudio build.

Returns a Utf8Error if the C *const char can’t be converted into a &’static str.

pub fn devices(&self) -> Result<Devices<'_>, Error>

Produces an iterator yielding the DeviceIndex for each device along with their respective DeviceInfos.

pub fn device_count(&self) -> Result<u32, Error>

Retrieve the number of available devices.

The number of available devices may be zero.

Returns an Error if PortAudio encounters some error.

NOTE: The only error documented by PortAudio that may occur calling this method is caused by PortAudio not being initialised, however this should not be possible with our type-safe PortAudio API.

pub fn default_input_device(&self) -> Result<DeviceIndex, Error>

Retrieve the index of the default input device. The result can be used in the InSettings used to open an In Stream.

Returns the default output device index for the default host API.

Returns Error if no default input device is available or an error was encountered.

TODO: Investigate exactly what errors may occur as the PA docs aren’t clear on this.

pub fn default_output_device(&self) -> Result<DeviceIndex, Error>

Retrieve the index of the default output device. The result can be used in the OutSettings used to open an Out Stream.

Returns the default input device index for the default host API.

Returns Error if no default input device is available or an error was encountered.

TODO: Investigate exactly what errors may occur as the PA docs aren’t clear on this.

pub fn device_info(&self, device: DeviceIndex) -> Result<DeviceInfo<'_>, Error>

Retrieve a DeviceInfo structure containing information about the specified device.

Returns Ok(DeviceInfo) if successful.

Returns Err(Error::InvalidDevice) if the device parameter is out of range.

Arguments

device - A valid device index in the range 0 to (port_audio.device_count()-1).

pub fn host_apis(&self) -> HostApis<'_>

Produces an iterator yielding the HostApiIndex of each available API along with their respective HostApiInfos.

pub fn host_api_count(&self) -> Result<HostApiIndex, Error>

Retrieve the number of available host APIs.

Even if a host API is available it may have no devices available.

Return a non-negative value indicating the number of available host APIs or an Error if an error is encountered.

TODO: Determine exactly what errors might occur (PA docs aren’t clear on this).

pub fn default_host_api(&self) -> Result<HostApiIndex, Error>

Retrieve the index of the default host API.

The default host API will be the lowest common denominator host API on the current platform and is unlikely to provide the best performance.

Return a non-negative value ranging from 0 to (get_host_api_count()-1) indicating the default host API index or an Error if an error is encountered.

TODO: Determine exactly what errors might occur (PA docs aren’t clear on this).

pub fn host_api_info<'a>( &'a self, host_api: HostApiIndex, ) -> Option<HostApiInfo<'a>>

Retrieve a pointer to a structure containing information about a specific host Api.

Arguments

host_api - A valid host API index ranging from 0 to (Pa_GetHostApiCount()-1)

Return Some(PaHostApiInfo) describing a specific host API.

Returns None if the host_api parameter is out of range or an error is encountered.

pub fn host_api_type_id_to_host_api_index( &self, type_id: HostApiTypeId, ) -> Result<HostApiIndex, Error>

Convert a static host API unique identifier, into a runtime host API index.

Arguments

typde_id - A unique host API identifier belonging to the PaHostApiTypeId enumeration.

Return a valid HostApiIndex ranging from 0 to (get_host_api_count()-1) or an Error if an error is encountered.

TODO: Determine exactly what errors might occur (PA docs aren’t clear on this).

pub fn api_device_index_to_device_index( &self, host_api: HostApiIndex, host_api_device_index: i32, ) -> Result<DeviceIndex, Error>

Convert a host-API-specific device index to standard PortAudio device index.

This function may be used in conjunction with the device_count field of HostApiInfo to enumerate all devices for the specified host API.

Arguments

host_api - A valid host API index ranging from 0 to (get_host_api_count()-1)

host_api_device_index - A valid per-host device index in the range 0 to (get_host_api_info(host_api).device_count-1)

Return a non-negative DeviceIndex ranging from 0 to (get_device_count()-1) or an Error if an error is encountered.

TODO: Determine exactly what errors might occur (PA docs aren’t clear on this).

pub fn is_input_format_supported<I>( &self, params: StreamParameters<I>, sample_rate: f64, ) -> Result<(), Error>
where I: Sample,

Determine whether it would be possible to open an input-only stream with the specified parameters.

The suggested_latency field of the StreamParameters is ignored.

Returns Ok(()) if the format is supported, and an Err(Error) indicating why the format is not supported otherwise.

pub fn is_output_format_supported<O>( &self, params: StreamParameters<O>, sample_rate: f64, ) -> Result<(), Error>
where O: Sample,

Determine whether it would be possible to open an output-only stream with the specified parameters.

The suggested_latency field of the StreamParameters is ignored.

Returns Ok(()) if the format is supported, and an Err(Error) indicating why the format is not supported otherwise.

pub fn is_duplex_format_supported<I, O>( &self, in_params: StreamParameters<I>, out_params: StreamParameters<O>, sample_rate: f64, ) -> Result<(), Error>
where I: Sample, O: Sample,

Determine whether it would be possible to open a duplex stream with the specified parameters.

The suggested_latency field of the StreamParameters is ignored.

Returns Ok(()) if the format is supported, and an Err(Error) indicating why the format is not supported otherwise.

pub fn open_blocking_stream<S>( &self, settings: S, ) -> Result<Stream<Blocking<<S::Flow as Flow>::Buffer>, S::Flow>, Error>
where S: StreamSettings, S::Flow: Flow,

Open a new blocking Stream with the given settings.

The Stream will be opened in Blocking “read/write” mode.

In Blocking mode, the client can receive sample data using Stream::read (for Input and Duplex streams only) and write sample data using Stream::write (for Output and Duplex streams only). The number of samples that may be read or written without blocking is returned by Stream::read_available and Stream::write_available respectively.

The returned Stream is inactive (stopped).

pub fn open_non_blocking_stream<S, C>( &self, settings: S, callback: C, ) -> Result<Stream<NonBlocking, S::Flow>, Error>
where S: StreamSettings, S::Flow: Flow, C: FnMut(<S::Flow as Flow>::CallbackArgs) -> PaStreamCallbackResult + 'static,

Open a new non-blocking Stream with the given settings.

When a non-blocking stream is running, PortAudio calls the given callback periodically. The callback function is responsible for processing buffers of audio samples passed via the input and/or output parameters (depending on the Stream’s Flow).

The callback runs at very high or real-time priority. It is required to consistently meet its time deadlines. Do not allocate memory, access the file system, call library functions or call other functions from the stream callback that may block or take an unpredictable amount of time to complete.

In order for a stream to maintain glitch-free operation the callback must consume and return audio data faster than it is recorded and/or played. PortAudio anticipates that each callback invocation may execute for a duration approaching the duration of frames audio frames at the stream sample rate. It is reasonable to expect to be able to utilise 70% or more of the available CPU time in the PortAudio callback. However, due to buffer size adaption and other factors, not all host APIs are able to guarantee audio stability under heavy CPU load with arbitrary fixed callback buffer sizes. When high callback CPU utilisation is required the most robust behaviour can be achieved by using an unspecified frames_per_buffer in the given settings (i.e. setting it to 0).

The callback should return one of the variants of the StreamCallbackResult enum. To ensure that the callback continues to be called, it should return Continue. Either Complete or Abort can be returned to finish stream processing, after either of these values is returned the callback will not be called again. If Abort is returned the stream will finish as soon as possible. If Complete is returned, the stream will continue until all buffers generated by the callback have been played. This may be useful in applications such as soundfile players where a specific duration of output is required. However, it is not necessary to utilize this mechanism as Stream::stop/abort/close can also be used to stop the Stream. Output stream callbacks must always fill the entire buffer irrespective of its return value.

The returned Stream is inactive (stopped).

pub fn default_input_stream_params<I>( &self, channels: i32, ) -> Result<StreamParameters<I>, Error>

Produce the default StreamParameters for an Input Stream.

The device used will be the default input device for the default Host API.

The produced Parameters will assume interleaved buffered audio data.

pub fn default_output_stream_params<O>( &self, channels: i32, ) -> Result<StreamParameters<O>, Error>

Produce the default StreamParameters for an Output Stream.

The device used will be the default output device for the default Host API.

The produced Parameters will assume interleaved buffered audio data.

pub fn default_input_stream_settings<I>( &self, channels: i32, sample_rate: f64, frames_per_buffer: u32, ) -> Result<InputStreamSettings<I>, Error>

Produce the default InputStreamSettings with the given number of channels, sample_rate and frames per buffer.

The device used will be the default input device for the default Host API.

The produced settings will assume interleaved buffered audio data.

pub fn default_output_stream_settings<O>( &self, channels: i32, sample_rate: f64, frames_per_buffer: u32, ) -> Result<OutputStreamSettings<O>, Error>

Produce the default OutputStreamSettings with the given number of channels, sample_rate and frames per buffer.

The device used will be the default output device for the default Host API.

The produced settings will assume interleaved buffered audio data.

pub fn default_duplex_stream_settings<I, O>( &self, in_channels: i32, out_channels: i32, sample_rate: f64, frames_per_buffer: u32, ) -> Result<DuplexStreamSettings<I, O>, Error>

Produce the default DuplexStreamSettings with the given number of channels, sample_rate and frames per buffer.

The devices used will be the default input and output devices for the default Host API.

The produced settings will assume interleaved buffered audio data.

pub fn sleep(&self, m_sec: i32)

Put the caller to sleep for at least ‘msec’ milliseconds.

In the original API this function is provided only as a convenience for authors of portable code (such as the tests and examples in the PortAudio distribution). This may be removed in the future in favour of rust’s ::std::thread::sleep function.

The function may sleep longer than requested so don’t rely on this for accurate musical timing.

pub fn last_host_error_info<'a>(&'a self) -> HostErrorInfo<'a>

Return information about the last host error encountered.

The error information returned by this method will never be modified asynchronously by errors occurring in other PortAudio owned threads (such as the thread that manages the stream callback).

This method is provided as a last resort, primarily to enhance debugging by providing clients with access to all available error information.

Return a pointer to an immutable structure constraining information about the host error. The values in this structure will only be valid if a PortAudio function or method has previously returned the UnanticipatedHostError error code.