Function portaudio::pa::get_sample_size

pub fn get_sample_size(format: PaSampleFormat) -> PaError

Retrieve the size of a given sample format in bytes.

Return the size in bytes of a single sample in the specified format, or PaSampleFormatNotSupported if the format is not supported.