Function portaudio::pa::host::api_device_index_to_device_index

pub fn api_device_index_to_device_index(host_api: PaHostApiIndex, host_api_device_index: int) -> PaDeviceIndex

Convert a host-API-specific device index to standard PortAudio device index. This function may be used in conjunction with the deviceCount field of PaHostApiInfo to enumerate all devices for the specified host API.

Arguments

host_api - A valid host API index ranging from 0 to (get_host_api_count()-1)

host_api_device_index - A valid per-host device index in the range 0 to (get_host_api_info(host_api).device_count-1)

Return a non-negative PaDeviceIndex ranging from 0 to (get_device_count()-1) or, a PaErrorCode (which are always negative) if PortAudio is not initialized or an error is encountered.