Module portaudio::stream

Contains the Stream class and associated values

Structs

Stream
An object for an PortAudio stream

StreamCallbackFlags
Flags indicating the status of the callback

StreamFlags
Flags used to control the behavior of a stream

StreamInfo
Information about the actual latency and sample rate values the stream uses

StreamParameters
Stream parameters to be used with Stream::open()

StreamTimeInfo
Time information for various stream related values

Enums

StreamCallbackResult
Allowable return values for a StreamCallback

Constants

pub const CLIP_OFF: StreamFlags = BitFlags { bits: $value }
Disable clipping of out of range samples

pub const DITHER_OFF: StreamFlags = BitFlags { bits: $value }
Disable dithering

pub const INPUT_OVERFLOW: StreamCallbackFlags = BitFlags { bits: $value }
Indicates that the callback has discarded some data

pub const INPUT_UNDERFLOW: StreamCallbackFlags = BitFlags { bits: $value }
Indicates that the callback has inserted one or more zeroes since not enough data was available

pub const NEVER_DROP_INPUT: StreamFlags = BitFlags { bits: $value }
Request that a full duplex stream will not discard overflowed input samples. The frames_per_buffer must be set to unspecified (0)

pub const OUTPUT_OVERFLOW: StreamCallbackFlags = BitFlags { bits: $value }
Indicates that certain data was discarded since there was no room

pub const OUTPUT_UNDERFLOW: StreamCallbackFlags = BitFlags { bits: $value }
Indicates that extra data was inserted in the output since there was not engough available

pub const PLATFORM_SPECIFIC: StreamFlags = BitFlags { bits: $value }
Range for platform specific flags. Not all of the upper 16 bits need to be set at the same time.

pub const PRIME_OUTPUT_BUFFERS_USING_STREAM_CALLBACK: StreamFlags = BitFlags { bits: $value }
Call the stream callback to fill initial output buffers, rather than priming the buffers with silence

pub const PRIMING_OUTPUT: StreamCallbackFlags = BitFlags { bits: $value }
Some or all of the output data will be used to prime the stream, input data may be zero

Statics

pub static FRAMES_PER_BUFFER_UNSPECIFIED: u64 = 0
Argument to Stream::open() or Stream::open_default() to allow PortAudio itself determine the optimal number of frames per buffer. This number may differ each time the callback is called.

Traits

SampleType
Types that are allowed to be used as samples in a Stream

Functions

is_format_supported
Returns Ok when the StreamParameters are supported. This ignores the latency field.

Type Definitions

StreamCallback
Callback to consume, process or generate audio

StreamFinishedCallback
Callback to be fired when a StreamCallback is stopped