Struct portaudio::types::PaDeviceInfo

pub struct PaDeviceInfo {
    pub struct_version: int,
    pub name: String,
    pub host_api: PaHostApiIndex,
    pub max_input_channels: int,
    pub max_output_channels: int,
    pub default_low_input_latency: PaTime,
    pub default_low_output_latency: PaTime,
    pub default_high_input_latency: PaTime,
    pub default_high_output_latency: PaTime,
    pub default_sample_rate: f64,
}

A structure providing information and capabilities of PortAudio devices. Devices may support input, output or both input and output.

Fields

struct_version

The version of the struct

name

The name of the devie

host_api

Host API identifier

max_input_channels

Maximal number of input channels for this device

max_output_channels

maximal number of output channel for this device

default_low_input_latency

The default low latency for input with this device

default_low_output_latency

The default low latency for output with this device

default_high_input_latency

The default high latency for input with this device

default_high_output_latency

The default high latency for output with this device

default_sample_rate

The default sample rate for this device